+++ components/Frame.js
@@ -102,9 +102,12 @@
       var _this = _possibleConstructorReturn(this, (Frame.__proto__ || Object.getPrototypeOf(Frame)).call(this, props));
 
       _this.state = {
+        dynamicView: { height: null, width: null },
         draggedLinkId: null,
         draggedItems: [],
+        offset: { x: 0, y: 0 },
         pointer: null,
+        scroll: { x: 0, y: 0 },
         showSelector: false,
         selectedItems: [],
         whenUpdated: getTime() // this attribute is used to force React render.
@@ -115,14 +118,37 @@
     _createClass(Frame, [{
       key: 'componentDidMount',
       value: function componentDidMount() {
+        var setState = this.setState.bind(this);
+
         var container = (0, _reactDom.findDOMNode)(this).parentNode;
+
+        window.addEventListener('scroll', function () {
+          setState({ scroll: {
+              x: window.scrollX,
+              y: window.scrollY
+            } });
+        });
+
+        window.addEventListener('resize', function () {
+          var rect = container.getBoundingClientRect();
+
+          setState({ dynamicView: {
+              height: rect.height,
+              width: rect.width
+            } });
+        });
 
         var offset = {
           x: container.offsetLeft,
           y: container.offsetTop
         };
 
-        this.setState({ offset: offset });
+        var scroll = {
+          x: window.scrollX,
+          y: window.scrollY
+        };
+
+        setState({ offset: offset, scroll: scroll });
       }
     }, {
       key: 'render',
@@ -153,14 +179,14 @@
         var _state = this.state,
             draggedItems = _state.draggedItems,
             draggedLinkId = _state.draggedLinkId,
-            offset = _state.offset,
             pointer = _state.pointer,
+            dynamicView = _state.dynamicView,
             selectedItems = _state.selectedItems,
             showSelector = _state.showSelector;
 
 
-        var height = view.height;
-        var width = view.width;
+        var height = dynamicView.height || view.height;
+        var width = dynamicView.width || view.width;
 
         var typeOfNode = item.util.typeOfNode;
 
@@ -170,9 +196,14 @@
         var setState = this.setState.bind(this);
 
         var getCoordinates = function getCoordinates(e) {
+          var _state2 = _this2.state,
+              offset = _state2.offset,
+              scroll = _state2.scroll;
+
+
           return {
-            x: e.clientX - offset.x,
-            y: e.clientY - offset.y
+            x: e.clientX - offset.x + scroll.x,
+            y: e.clientY - offset.y + scroll.y
           };
         };
 
@@ -606,10 +637,8 @@
     style: { border: '1px solid black' },
     updateLink: Function.prototype,
     view: {
-      height: 400,
       link: {},
-      node: {},
-      width: 400
+      node: {}
     }
   };